Harden sql database
We cannot expect that each IO operation or sql transactions
will succeed. Thus the class needs to deal with this ambiguity.
This CL hardens the sql_database.h.

Test: "passes unittests. Works on DUT."
Bug: b:292164625
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4707383
Commit-Queue: Dmitry Grebenyuk <dgrebenyuk@google.com>
Cr-Commit-Position: refs/heads/main@{#1174041}

+++ chrome/browser/ash/app_list/search/local_image_search/annotation_storage.cc
@@ -34,6 +34,9 @@ constexpr int kVersionNumber = 3;
 // The table cannot exist when calling this function.
 int CreateNewSchema(SqlDatabase* db) {
   DVLOG(1) << "Making a table";
+  if (!db) {
+    return 0;
+  }
 
   static constexpr char kQuery[] =
       // clang-format off
@@ -43,24 +46,27 @@ int CreateNewSchema(SqlDatabase* db) {
           "last_modified_time INTEGER NOT NULL,"
           "is_ignored INTEGER NOT NULL)";
   // clang-format on
-  sql::Statement statement = db->GetStatementForQuery(SQL_FROM_HERE, kQuery);
-  if (!statement.Run()) {
+  std::unique_ptr<sql::Statement> statement =
+      db->GetStatementForQuery(SQL_FROM_HERE, kQuery);
+  if (!statement || !statement->Run()) {
     return 0;
   }
 
   static constexpr char kQuery1[] =
       "CREATE INDEX ind_annotations_label ON annotations(label)";
 
-  sql::Statement statement1 = db->GetStatementForQuery(SQL_FROM_HERE, kQuery1);
-  if (!statement1.Run()) {
+  std::unique_ptr<sql::Statement> statement1 =
+      db->GetStatementForQuery(SQL_FROM_HERE, kQuery1);
+  if (!statement1 || !statement1->Run()) {
     return 0;
   }
 
   static constexpr char kQuery2[] =
       "CREATE INDEX ind_annotations_image_path ON annotations(image_path)";
 
-  sql::Statement statement2 = db->GetStatementForQuery(SQL_FROM_HERE, kQuery2);
-  if (!statement2.Run()) {
+  std::unique_ptr<sql::Statement> statement2 =
+      db->GetStatementForQuery(SQL_FROM_HERE, kQuery2);
+  if (!statement2 || !statement2->Run()) {
     return 0;
   }
 
@@ -68,13 +74,18 @@ int CreateNewSchema(SqlDatabase* db) {
 }
 
 int MigrateSchema(SqlDatabase* db, int current_version_number) {
+  if (!db) {
+    return 0;
+  }
+
   if (current_version_number == kVersionNumber) {
     return current_version_number;
   }
 
   static constexpr char kQuery[] = "DROP TABLE IF EXISTS annotations";
-  sql::Statement statement = db->GetStatementForQuery(SQL_FROM_HERE, kQuery);
-  if (!statement.Run()) {
+  std::unique_ptr<sql::Statement> statement =
+      db->GetStatementForQuery(SQL_FROM_HERE, kQuery);
+  if (!statement || !statement->Run()) {
     return 0;
   }
 
@@ -154,15 +165,18 @@ void AnnotationStorage::Insert(const ImageInfo& image_info) {
   // clang-format on
 
   for (const auto& annotation : image_info.annotations) {
-    sql::Statement statement =
+    std::unique_ptr<sql::Statement> statement =
         sql_database_->GetStatementForQuery(SQL_FROM_HERE, kQuery);
+    if (!statement) {
+      return;
+    }
     DVLOG(1) << annotation;
-    statement.BindString(0, annotation);
-    statement.BindString(1, image_info.path.value());
-    statement.BindTime(2, image_info.last_modified);
-    statement.BindInt(3, image_info.is_ignored);
+    statement->BindString(0, annotation);
+    statement->BindString(1, image_info.path.value());
+    statement->BindTime(2, image_info.last_modified);
+    statement->BindInt(3, image_info.is_ignored);
 
-    if (!statement.Run()) {
+    if (!statement->Run()) {
       // TODO(b/260646344): log to UMA instead.
       return;
     }
@@ -176,11 +190,15 @@ void AnnotationStorage::Remove(const base::FilePath& image_path) {
 
   static constexpr char kQuery[] = "DELETE FROM annotations WHERE image_path=?";
 
-  sql::Statement statement =
+  std::unique_ptr<sql::Statement> statement =
       sql_database_->GetStatementForQuery(SQL_FROM_HERE, kQuery);
-  statement.BindString(0, image_path.value());
+  if (!statement) {
+    return;
+  }
 
-  statement.Run();
+  statement->BindString(0, image_path.value());
+
+  statement->Run();
 }
 
 std::vector<ImageInfo> AnnotationStorage::GetAllAnnotations() {
@@ -194,17 +212,20 @@ std::vector<ImageInfo> AnnotationStorage::GetAllAnnotations() {
           "ORDER BY label";
   // clang-format on
 
-  sql::Statement statement =
+  std::unique_ptr<sql::Statement> statement =
       sql_database_->GetStatementForQuery(SQL_FROM_HERE, kQuery);
+  if (!statement) {
+    return {};
+  }
 
   std::vector<ImageInfo> matched_paths;
-  while (statement.Step()) {
-    const base::FilePath path = base::FilePath(statement.ColumnString(1));
-    const base::Time time = statement.ColumnTime(2);
-    const bool is_ignored = statement.ColumnBool(3);
-    DVLOG(1) << "Select find: " << statement.ColumnString(0) << ", " << path
+  while (statement->Step()) {
+    const base::FilePath path = base::FilePath(statement->ColumnString(1));
+    const base::Time time = statement->ColumnTime(2);
+    const bool is_ignored = statement->ColumnBool(3);
+    DVLOG(1) << "Select find: " << statement->ColumnString(0) << ", " << path
              << ", " << time;
-    matched_paths.push_back({{statement.ColumnString(0)},
+    matched_paths.push_back({{statement->ColumnString(0)},
                              std::move(path),
                              std::move(time),
                              is_ignored});
@@ -227,18 +248,21 @@ std::vector<ImageInfo> AnnotationStorage::FindImagePath(
           "ORDER BY label";
   // clang-format on
 
-  sql::Statement statement =
+  std::unique_ptr<sql::Statement> statement =
       sql_database_->GetStatementForQuery(SQL_FROM_HERE, kQuery);
-  statement.BindString(0, image_path.value());
+  if (!statement) {
+    return {};
+  }
+  statement->BindString(0, image_path.value());
 
   std::vector<ImageInfo> matched_paths;
-  while (statement.Step()) {
-    const base::FilePath path = base::FilePath(statement.ColumnString(1));
-    const base::Time time = statement.ColumnTime(2);
-    const bool is_ignored = statement.ColumnBool(3);
-    DVLOG(1) << "Select find: " << statement.ColumnString(0) << ", " << path
+  while (statement->Step()) {
+    const base::FilePath path = base::FilePath(statement->ColumnString(1));
+    const base::Time time = statement->ColumnTime(2);
+    const bool is_ignored = statement->ColumnBool(3);
+    DVLOG(1) << "Select find: " << statement->ColumnString(0) << ", " << path
              << ", " << time;
-    matched_paths.push_back({{statement.ColumnString(0)},
+    matched_paths.push_back({{statement->ColumnString(0)},
                              std::move(path),
                              std::move(time),
                              is_ignored});
@@ -263,25 +287,28 @@ std::vector<FileSearchResult> AnnotationStorage::Search(
           "ORDER BY image_path";
   // clang-format on
 
-  sql::Statement statement =
+  std::unique_ptr<sql::Statement> statement =
       sql_database_->GetStatementForQuery(SQL_FROM_HERE, kQuery);
-  statement.BindString(0, base::StrCat({"%", base::UTF16ToUTF8(query), "%"}));
+  if (!statement) {
+    return {};
+  }
+  statement->BindString(0, base::StrCat({"%", base::UTF16ToUTF8(query), "%"}));
 
   std::vector<FileSearchResult> matched_paths;
   TokenizedString tokenized_query(query);
   ash::string_matching::FuzzyTokenizedStringMatch fuzzy_match;
-  while (statement.Step()) {
+  while (statement->Step()) {
     double relevance = fuzzy_match.Relevance(
         tokenized_query,
-        TokenizedString(base::UTF8ToUTF16(statement.ColumnString(0))),
+        TokenizedString(base::UTF8ToUTF16(statement->ColumnString(0))),
         /*use_weighted_ratio=*/true);
     if (relevance < kRelevanceThreshold) {
       continue;
     }
 
-    const base::FilePath path = base::FilePath(statement.ColumnString(1));
-    const base::Time time = statement.ColumnTime(2);
-    DVLOG(1) << "Select: " << statement.ColumnString(0) << ", " << path << ", "
+    const base::FilePath path = base::FilePath(statement->ColumnString(1));
+    const base::Time time = statement->ColumnTime(2);
+    DVLOG(1) << "Select: " << statement->ColumnString(0) << ", " << path << ", "
              << time << " rl: " << relevance;
 
     if (matched_paths.empty() || matched_paths.back().file_path != path) {
